updated the products table schema

+++ schema.sql
@@ -10,7 +10,7 @@ product_name VARCHAR(100) NOT NULL,
 department_name VARCHAR(50) NOT NULL,
 price DECIMAL(10,2) NOT NULL,
 stock_quantity INT NOT NULL,
-product_sales INT NOT NULL,
+product_sales DECIMAL(10,2) NOT NULL,
 PRIMARY KEY(item_id)
 );
